TimeUtils moved into /java/ package from /bean/common/ package and PharmaceuticalBillItem remove its TimeUtils usage due to not being used
